Replace existing harvester/infomedia module with harvester/retriever
This also entails a new dataio-harvester-retriever docker image.
