--- harvester/infomedia/src/main/docker/harvester-infomedia.json
@@ -1,4 +1,0 @@
-{
-  "appName": "InfomediaHarvester",
-  "profiles": ["dmz-proxy"]
-}
