--- harvester/infomedia/src/main/java/dk/dbc/dataio/harvester/infomedia/JobSpecificationTemplate.java
@@ -1,28 +1,0 @@
-package dk.dbc.dataio.harvester.infomedia;
-
-import dk.dbc.dataio.commons.types.JobSpecification;
-import dk.dbc.dataio.harvester.types.HarvesterException;
-import dk.dbc.dataio.harvester.types.InfomediaHarvesterConfig;
-
-class JobSpecificationTemplate {
-    public static final int SUBMITTER_NUMBER = 190002;
-
-    static JobSpecification create(InfomediaHarvesterConfig config) throws HarvesterException {
-        try {
-            final InfomediaHarvesterConfig.Content configFields = config.getContent();
-            return new JobSpecification()
-                    .withPackaging("addi-json")
-                    .withFormat(configFields.getFormat())
-                    .withCharset("utf8")
-                    .withDestination(configFields.getDestination())
-                    .withSubmitterId(SUBMITTER_NUMBER)
-                    .withMailForNotificationAboutVerification("placeholder")
-                    .withMailForNotificationAboutProcessing("placeholder")
-                    .withResultmailInitials("placeholder")
-                    .withDataFile("placeholder")
-                    .withType(JobSpecification.Type.INFOMEDIA);
-        } catch (RuntimeException e) {
-            throw new HarvesterException("Unable to create job specification template", e);
-        }
-    }
-}
