--- harvester/infomedia/src/main/java/dk/dbc/dataio/harvester/infomedia/rest/StatusBean.java
@@ -1,10 +1,0 @@
-package dk.dbc.dataio.harvester.infomedia.rest;
-
-import dk.dbc.dataio.commons.utils.service.ServiceStatus;
-import jakarta.ejb.Stateless;
-import jakarta.ws.rs.Path;
-
-@Stateless
-@Path("/")
-public class StatusBean implements ServiceStatus {
-}
