+++ harvester/retriever/src/main/docker/Dockerfile
