+++ harvester/retriever/src/main/docker/harvester-retriever.json
@@ -1,0 +1,4 @@
+{
+  "appName": "RetrieverHarvester",
+  "profiles": ["dmz-proxy"]
+}
