+++ harvester/retriever/src/main/java/dk/dbc/dataio/harvester/retriever/ArticlePayload.java
@@ -1,0 +1,40 @@
+package dk.dbc.dataio.harvester.retriever;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import dk.dbc.dataio.commons.creatordetector.connector.CreatorNameSuggestion;
+import dk.dbc.retriever.connector.model.Article;
+import dk.dbc.tagstack.connector.model.TagResult;
+
+import java.util.List;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class ArticlePayload {
+
+    private Article article;
+    private List<CreatorNameSuggestion> creatorNameSuggestions;
+    private List<TagResult> tags;
+
+    public Article getArticle() {
+        return article;
+    }
+
+    public void setArticle(Article article) {
+        this.article = article;
+    }
+
+    public List<CreatorNameSuggestion> getCreatorNameSuggestions() {
+        return creatorNameSuggestions;
+    }
+
+    public void setCreatorNameSuggestions(List<CreatorNameSuggestion> creatorNameSuggestions) {
+        this.creatorNameSuggestions = creatorNameSuggestions;
+    }
+
+    public List<TagResult> getTags() {
+        return tags;
+    }
+
+    public void setTags(List<TagResult> tags) {
+        this.tags = tags;
+    }
+}
