+++ harvester/retriever/src/main/java/dk/dbc/dataio/harvester/retriever/rest/StatusBean.java
@@ -1,0 +1,10 @@
+package dk.dbc.dataio.harvester.retriever.rest;
+
+import dk.dbc.dataio.commons.utils.service.ServiceStatus;
+import jakarta.ejb.Stateless;
+import jakarta.ws.rs.Path;
+
+@Stateless
+@Path("/")
+public class StatusBean implements ServiceStatus {
+}
