+++ pom.xml
@@ -617,17 +617,6 @@
             </dependency>
             <dependency>
                 <groupId>dk.dbc</groupId>
-                <artifactId>infomedia-connector</artifactId>
-                <version>2.0-SNAPSHOT</version>
-                <exclusions>
-                    <exclusion>
-                        <groupId>dk.dbc</groupId>
-                        <artifactId>dbc-commons-httpclient</artifactId>
-                    </exclusion>
-                </exclusions>
-            </dependency>
-            <dependency>
-                <groupId>dk.dbc</groupId>
                 <artifactId>invariant-utils</artifactId>
                 <version>0.1.0</version>
             </dependency>
